+++ aio_wx_widgets/panels/panel.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import logging
 
 import wx
-
+from wx.lib.scrolledpanel import ScrolledPanel
 from aio_wx_widgets import type_annotations as T
 from aio_wx_widgets.sizers import PanelMixin, SizerMixin
 
@@ -14,9 +14,14 @@ _LOGGER = logging.getLogger(__name__)
 class SimplePanel(PanelMixin, SizerMixin):
     """A simple panel."""
 
-    def __init__(self, parent, **kwargs):
+    def __init__(self, parent, scrollable=False, **kwargs):
         """Init."""
-        self.ui_item = wx.Panel(parent)
+
+        if scrollable:
+            self.ui_item = ScrolledPanel(parent)
+            self.ui_item.SetupScrolling()
+        else:
+            self.ui_item = wx.Panel(parent)
         super().__init__()
 
         self._sizer = wx.BoxSizer(wx.VERTICAL)