+++ aio_wx_widgets/widgets/group.py
@@ -6,6 +6,7 @@ import logging
 import wx
 
 from aio_wx_widgets.sizers import SizerMixin
+from aio_wx_widgets.widgets import text
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -65,3 +66,36 @@ class Group(SizerMixin):
 
     def _finish(self):
         self._sizer.AddSpacer(self._bottom_border)
+
+
+class Section(SizerMixin):
+    """A section with a large header text and an optional closing line.
+
+    LARGE HEADER TEXT
+
+    - content
+    - content
+    ----------------------------------- [optional closing line]
+    """
+
+    def __init__(self, header: str, add_closing_line=True):
+        """Init."""
+        self.ui_item = self._sizer = wx.BoxSizer(orient=wx.VERTICAL)
+        self._parent = None
+        self._header = header
+        self._add_closing_line = add_closing_line
+
+    def __call__(self, parent):
+        self._parent = parent
+        return self
+
+    def __enter__(self):
+        """Enter the context manager."""
+        self.add(
+            text.Text(text=self._header, font_size=1.5, bold=True), margin=(5, 0, 15, 0)
+        )
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Exit the context manager"""
+        self.add(wx.StaticLine(self._parent), create=False, margin=(0, 0, 15, 1))