+++ aio_wx_widgets/widgets/validators/__init__.py
@@ -1,3 +1,4 @@
 """All kinds of validators."""
 
 from aio_wx_widgets.widgets.validators.validators import *  # noqa: F403,F401
+from aio_wx_widgets.widgets.validators.exception import *  # noqa: F403,F401