+++ aio_wx_widgets/widgets/validators/exception.py
@@ -1,0 +1,7 @@
+"""Validation exception"""
+
+__all__ = ["ValidationError"]
+
+
+class ValidationError(Exception):
+    """Input validation error"""