+++ demo/controller/controller_one.py
@@ -11,8 +11,10 @@ _LOGGER = logging.getLogger(__name__)
 
 class ControllerOne(BaseController):
     def __init__(self, model):
-        self.value_1: int = 0
+        self.float_val: float = 0
+        self.int_val: int = 0
         self.a_string_value = "A certain string"
+        self.a_checkbox_value: bool = True
         super().__init__(model)
 
         self.add_to_log = Events()