+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.0.9
+
+- Add "Section widget"
+- Add better validation
+- Add Events package for pub sub inside controller-view part.
+- Add split/unsplit to splitview.
+- Add Checkbox
+
 ## v0.0.8
 
 - Add basic validation feedback.